% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{test_func}
\alias{test_func}
\title{A Cat Function}
\usage{
test_func(x = c(1, 2, 3))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
